refactor disease module to use DiseaseCacheRepository
