+++ agr_java_core/src/main/java/org/alliancegenome/core/service/DiseaseAnnotationFiltering.java
@@ -1,47 +1,47 @@
-package org.alliancegenome.api.service;
+package org.alliancegenome.core.service;
 
 import org.alliancegenome.es.model.query.FieldFilter;
 import org.alliancegenome.neo4j.entity.DiseaseAnnotation;
 import org.alliancegenome.neo4j.entity.node.Gene;
 
-import java.util.*;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import java.util.stream.Collectors;
-
-import static org.alliancegenome.api.service.FilterFunction.contains;
-import static org.alliancegenome.api.service.FilterFunction.fullMatchMultiValueOR;
 
 public class DiseaseAnnotationFiltering {
 
 
     public static FilterFunction<DiseaseAnnotation, String> termNameFilter =
-            (annotation, value) -> contains(annotation.getDisease().getName(), value);
+            (annotation, value) -> FilterFunction.contains(annotation.getDisease().getName(), value);
 
     public static FilterFunction<DiseaseAnnotation, String> geneticEntityFilter =
             (annotation, value) -> {
                 if (annotation.getFeature() == null)
                     return false;
-                return contains(annotation.getFeature().getSymbolText(), value);
+                return FilterFunction.contains(annotation.getFeature().getSymbolText(), value);
             };
 
     public static FilterFunction<DiseaseAnnotation, String> associationFilter =
-            (annotation, value) -> contains(annotation.getAssociationType(), value);
+            (annotation, value) -> FilterFunction.contains(annotation.getAssociationType(), value);
 
     public static FilterFunction<DiseaseAnnotation, String> sourceFilter =
-            (annotation, value) -> contains(annotation.getSource().getName(), value);
+            (annotation, value) -> FilterFunction.contains(annotation.getSource().getName(), value);
 
     public static FilterFunction<DiseaseAnnotation, String> geneticEntityTypeFilter =
             (annotation, value) -> FilterFunction.fullMatchMultiValueOR(annotation.getGeneticEntityType(), value);
 
     public static FilterFunction<DiseaseAnnotation, String> geneNameFilter =
-            (annotation, value) -> contains(annotation.getGene().getSymbol(), value);
+            (annotation, value) -> FilterFunction.contains(annotation.getGene().getSymbol(), value);
 
     public static FilterFunction<DiseaseAnnotation, String> geneSpeciesFilter =
-            (annotation, value) -> contains(annotation.getGene().getSpecies().getName(), value);
+            (annotation, value) -> FilterFunction.contains(annotation.getGene().getSpecies().getName(), value);
 
     public static FilterFunction<DiseaseAnnotation, String> evidenceCodeFilter =
             (annotation, value) -> {
                 Set<Boolean> filteringPassed = annotation.getEvidenceCodes().stream()
-                        .map(evidenceCode -> contains(evidenceCode.getName(), value))
+                        .map(evidenceCode -> FilterFunction.contains(evidenceCode.getName(), value))
                         .collect(Collectors.toSet());
                 return !filteringPassed.contains(false);
             };
@@ -49,7 +49,7 @@ public class DiseaseAnnotationFiltering {
     public static FilterFunction<DiseaseAnnotation, String> referenceFilter =
             (annotation, value) -> {
                 Set<Boolean> filteringPassed = annotation.getPublications().stream()
-                        .map(publication -> contains(publication.getPubId(), value))
+                        .map(publication -> FilterFunction.contains(publication.getPubId(), value))
                         .collect(Collectors.toSet());
                 // return true if at least one pub is found
                 return filteringPassed.contains(true);
@@ -60,7 +60,7 @@ public class DiseaseAnnotationFiltering {
                 Gene orthologyGene = annotation.getOrthologyGene();
                 if (orthologyGene == null)
                     return false;
-                return contains(orthologyGene.getSymbol(), value);
+                return FilterFunction.contains(orthologyGene.getSymbol(), value);
             };
 
     public static FilterFunction<DiseaseAnnotation, String> orthologSpeciesFilter =
@@ -68,7 +68,7 @@ public class DiseaseAnnotationFiltering {
                 Gene orthologyGene = annotation.getOrthologyGene();
                 if (orthologyGene == null)
                     return false;
-                return contains(orthologyGene.getSpecies().getName(), value);
+                return FilterFunction.contains(orthologyGene.getSpecies().getName(), value);
             };
