+++ agr_java_core/src/main/java/org/alliancegenome/core/service/PaginationResult.java
@@ -1,4 +1,4 @@
-package org.alliancegenome.api.service;
+package org.alliancegenome.core.service;
 
 import lombok.Getter;
 import lombok.Setter;
